chore: 캐릭터 추가
