+++ src/routes/AppRoutes.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from '../containers/Home.tsx';
 import Layout from '../components/Layout.tsx';
 import Lobby from '../containers/Lobby.tsx';
 import Test from '../containers/Test.tsx';
+import Home from '../containers/Home/Home.tsx';
+
 const AppRoute = () => {
     return(
     <Router>
