+++ src/styles/box.scss
@@ -2,7 +2,6 @@ $blue1-color: #BAC1FF;
 $blue2-color: #50578F;
 
 
-
 .box {
     background-color: $blue1-color;
     width: 300px;
